tokens: refactor - update react files structure
