+++ src/scripts/modules/tokens/react/Index/CreateTokenModal.jsx
@@ -1,9 +1,9 @@
 import React, {PropTypes} from 'react';
 import {Map} from 'immutable';
 import {Modal} from 'react-bootstrap';
-import ConfirmButtons from '../../../react/common/ConfirmButtons';
+import ConfirmButtons from '../../../../react/common/ConfirmButtons';
 import TokenString from './TokenString';
-import TokenEditor from './TokenEditor';
+import TokenEditor from '../tokenEditor/TokenEditor';
 
 export default React.createClass({
