+++ src/scripts/modules/tokens/react/Index/CreatedDate.jsx
