+++ src/scripts/modules/tokens/react/TokenDetail.jsx
@@ -3,7 +3,7 @@ import {Link} from 'react-router';
 import ApplicationStore from '../../../stores/ApplicationStore';
 import TokensStore from '../StorageTokensStore';
 import createStoreMixin from '../../../react/mixins/createStoreMixin';
-import TokenEditor from './TokenEditor';
+import TokenEditor from './tokenEditor/TokenEditor';
 import TokensActions from '../actionCreators';
 import {Map} from 'immutable';
 import BucketsStore from '../../components/stores/StorageBucketsStore';
